moved resolving code from view to middleware, no need for urlconf entry
